+++ src/lib/andromeda/Utilities.hpp
@@ -135,7 +135,7 @@ public:
      * @param bytes number of bytes to print
      * @param width number of bytes per line
      */
-    void DumpBytes(const void* ptr, size_t bytes, size_t width = 16);
+    void DumpBytes(const void* ptr, uint64_t bytes, uint8_t width = 16);
 
     /** Append to an internal buffer that can be shown with an empty Print */
     template <class T> Debug& operator<<(const T& dat)
